Divided more components and data types
