+++ app/types/categoryTypes.ts
@@ -1,0 +1,4 @@
+export interface Category {
+  id_cat: number;
+  name: string;
+}
